Add files via upload
The Memory Game is a graphical game implemented in Python using the Pygame library. The objective of the game is to match pairs of cards by flipping them over one at a time.

The game starts by displaying a grid of cards on the screen, faced down. The player can click on two cards to reveal their images. If the two cards have the same image, they are considered a match and remain face up. Otherwise, the cards are flipped back faced down. The player continues to flip cards until all pairs are matched.
.
To run this code, make sure you have Pygame installed (pip install pygame). The code assumes that you have card images named card1.png, card2.png, etc., in the same directory as the script. You can replace these images with your own card designs.

This implementation allows the player to click on two cards at a time and tries to match them.